requirements.txt
yapf==0.32.0
GitPython==3.1.31
tqdm==4.65.0
mypy==1.1.1
matplotlib==3.7.1
requests==2.28.2
libcst==1.0.0
PyDriller==2.5
pandas==2.0.3
patsy==0.5.3
statsmodels==0.14.0